Move timer logic into Timer component
